refactor: 중복되는 코드 함수화
- 개선: 이를 dataMutation 함수를 이용해 분리하고 useMachines, useFoods에서 인자를 전달해 처리하도록 분리했습니다.
