+++ src/hooks/useProducts.jsx
@@ -1,93 +1,11 @@
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import { useQuery } from "@tanstack/react-query";
 import {
-  getMachines,
-  setCustomMachines,
-  getFoods,
-  setCustomFoods,
   getCustomMachines,
   submitMachines,
   getCustomFoods,
   submitFoods,
 } from "../api/Products";
 import { useState } from "react";
-
-export function useMachines() {
-  const queryClient = useQueryClient();
-  const [products, setProducts] = useState([]);
-  const [success, setSuccess] = useState();
-
-  const productsQuery = useQuery(["machines"], getMachines, {
-    staleTime: Infinity,
-    cacheTime: Infinity,
-  });
-
-  const addCustomMachines = useMutation(
-    ({ products }) => setCustomMachines(products),
-    {
-      onSuccess: () => {
-        queryClient.invalidateQueries(["customMachines"]);
-        queryClient.invalidateQueries(["customMachinesTemp"]);
-        queryClient.invalidateQueries(["machines"]);
-      },
-    }
-  );
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    addCustomMachines.mutate(
-      { products },
-      {
-        onSuccess: () => {
-          setSuccess(true);
-          setTimeout(() => {
-            setSuccess(null);
-          }, 4000);
-        },
-      }
-    );
-  };
-
-  return { productsQuery, handleSubmit, success, setProducts };
-}
-
-export function useFoods() {
-  const [success, setSuccess] = useState();
-  const [products, setProducts] = useState([]);
-  const queryClient = useQueryClient();
-
-  const productsQuery = useQuery(["foods"], getFoods, {
-    staleTime: Infinity,
-    cacheTime: Infinity,
-  });
-
-  const addCustomFoods = useMutation(
-    ({ products }) => setCustomFoods(products),
-    {
-      onSuccess: () => {
-        queryClient.invalidateQueries(["customFoods"]);
-        queryClient.invalidateQueries(["customFoodsTemp"]);
-        queryClient.invalidateQueries(["foods"]);
-      },
-    }
-  );
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    addCustomFoods.mutate(
-      { products },
-      {
-        onSuccess: () => {
-          setSuccess(true);
-          setTimeout(() => {
-            setSuccess(null);
-          }, 4000);
-        },
-      }
-    );
-  };
-
-  return { productsQuery, setProducts, success, handleSubmit };
-}
 
 export function useCustomProducts({ location, setProductsTemp }) {
   const [selectManager, setSelectManager] = useState("");
@@ -156,7 +74,7 @@ export function useCustomMachines() {
 }
 
 export function useCustomFoods() {
-  const productsQuery = useQuery(["customFoods"], getCustomFoods, {
+  const praoductsQuery = useQuery(["customFoods"], getCustomFoods, {
     staleTime: Infinity,
     cacheTime: Infinity,
   });
@@ -165,7 +83,7 @@ export function useCustomFoods() {
     submitFoods({
       mgrname: selectManager?.label,
       customFood: products,
-      time: location?.state,
+      time: location?.stte,
     });
 
   return { productsQuery, setProductsTemp };
